+++ SimPEG/electromagnetics/frequency_domain_1d/__init__.py
@@ -19,7 +19,7 @@ from . import supporting_functions
 #from .regularization import (
 #    LateralConstraint, get_2d_mesh
 #)
-from .IO import ModelIO
-import os
-import glob
-import unittest
+#from .IO import ModelIO
+#import os
+#import glob
+#import unittest
